+++ README.md
@@ -1,1 +1,3 @@
 # jwp-dashboard-jdbc
+## 🚀 1단계 - JDBC 라이브러리 구현하기
+- [x] DataSourcePopulatorUtils jdbc 모듈로 이동 + 리팩터링
