+++ app/src/main/java/com/techcourse/config/DataSourceConfig.java
@@ -2,11 +2,12 @@ package com.techcourse.config;
 
 import org.h2.jdbcx.JdbcDataSource;
 
+import javax.sql.DataSource;
 import java.util.Objects;
 
 public class DataSourceConfig {
 
-    private static javax.sql.DataSource INSTANCE;
+    private static DataSource INSTANCE;
 
     public static javax.sql.DataSource getInstance() {
         if (Objects.isNull(INSTANCE)) {
