+++ app/src/main/java/com/techcourse/support/context/ContextLoaderListener.java
@@ -1,10 +1,10 @@
 package com.techcourse.support.context;
 
 import com.techcourse.config.DataSourceConfig;
-import com.techcourse.support.jdbc.init.DatabasePopulatorUtils;
 import jakarta.servlet.ServletContextEvent;
 import jakarta.servlet.ServletContextListener;
 import jakarta.servlet.annotation.WebListener;
+import nextstep.datasource.DatabasePopulatorUtils;
 
 @WebListener
 public class ContextLoaderListener implements ServletContextListener {
