--- app/src/main/java/com/techcourse/support/jdbc/init/DatabasePopulatorUtils.java
@@ -1,47 +1,0 @@
-package com.techcourse.support.jdbc.init;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import javax.sql.DataSource;
-import java.io.File;
-import java.io.IOException;
-import java.net.URL;
-import java.nio.file.Files;
-import java.sql.Connection;
-import java.sql.SQLException;
-import java.sql.Statement;
-
-public class DatabasePopulatorUtils {
-
-    private static final Logger log = LoggerFactory.getLogger(DatabasePopulatorUtils.class);
-
-    public static void execute(DataSource dataSource) {
-        Connection connection = null;
-        Statement statement = null;
-        try {
-            final URL url = DatabasePopulatorUtils.class.getClassLoader().getResource("schema.sql");
-            final File file = new File(url.getFile());
-            final String sql = Files.readString(file.toPath());
-            connection = dataSource.getConnection();
-            statement = connection.createStatement();
-            statement.execute(sql);
-        } catch (NullPointerException | IOException | SQLException e) {
-            log.error(e.getMessage(), e);
-        } finally {
-            try {
-                if (statement != null) {
-                    statement.close();
-                }
-            } catch (SQLException ignored) {}
-
-            try {
-                if (connection != null) {
-                    connection.close();
-                }
-            } catch (SQLException ignored) {}
-        }
-    }
-
-    private DatabasePopulatorUtils() {}
-}
